Script modification
AnimationController was refactor, now it can be use in any object that would like to be animated in the future. 1. Drag the script to the object you would like to animate 2. Add the component animator to the object 3. In the script were you'll control the movement of the object make an instance of the AnimationController script and get the component of the AnimationController in the Start method 4. Use the ChangeAnimationStateTo of the Animation script to pass the animation you'll want to use in the object. 5. For better readability add a constant at the beginning of the object MovementSript of type string with the name of the animation to pass to the ChangeAnimationStateTo method.

EnemyMovement was refactor for better readability and scability